+++ mmdet/cv_core/utils/__init__.py
@@ -16,3 +16,4 @@ from .logging import print_log, get_logger
 from .receptivefield import calc_receptive_filed
 from .kmean import Kmean
 from .featuremap_vis import FeatureMapVis
+from .coco_creator import CocoCreator