+++ mmdet/datasets/__init__.py
@@ -7,10 +7,12 @@ from .voc import VOCDataset
 from .wider_face import WIDERFaceDataset
 from .xml_style import XMLDataset
 from .utils import replace_ImageToTensor
+from .voc_cocoformat import VOC_CocoFormat
+from .widerface_cocoformat import WiderFace_CocoFormat
 
 __all__ = [
-    'CustomDataset', 'XMLDataset', 'CocoDataset',
-    'VOCDataset', 'GroupSampler',
+    'CustomDataset', 'XMLDataset', 'CocoDataset', 'VOC_CocoFormat',
+    'VOCDataset', 'GroupSampler', 'WiderFace_CocoFormat',
     'build_dataloader', 'ConcatDataset', 'RepeatDataset', 'WIDERFaceDataset', 'DATASETS', 'PIPELINES',
     'build_dataset', 'replace_ImageToTensor'
 ]