+++ mmdet/datasets/voc_cocoformat.py
@@ -1,0 +1,8 @@
+from .coco import CocoDataset
+
+
+class VOC_CocoFormat(CocoDataset):
+    CLASSES = ('aeroplane', 'bicycle', 'bird', 'boat', 'bottle', 'bus', 'car',
+               'cat', 'chair', 'cow', 'diningtable', 'dog', 'horse',
+               'motorbike', 'person', 'pottedplant', 'sheep', 'sofa', 'train',
+               'tvmonitor')