+++ mmdet/datasets/widerface_cocoformat.py
@@ -1,0 +1,5 @@
+from .coco import CocoDataset
+
+
+class WiderFace_CocoFormat(CocoDataset):
+    CLASSES = ('face', )